Bug 834 	  	Integrate changes from Pde optimization for Finite Volume C++ library
svn path=/vcell/trunk/numerics/; revision=2569


Former-commit-id: 0c521038dd9532afd8a77b5f6a725590f8c5d6f8

+++ ExpressionParser/ASTMultNode.cpp
@@ -39,7 +39,7 @@ string ASTMultNode::infixString(int lang, NameScope* nameScope)
 
 void ASTMultNode::getStackElements(vector<StackElement>& elements) {
 	
-	int startSize = elements.size();
+	int startSize = (int)elements.size();
 	
 	//
 	// Add the stack machine instructions for each boolean child (followed by a branch "BZ" with unknown offset).
@@ -81,7 +81,7 @@ void ASTMultNode::getStackElements(vector<StackElement>& elements) {
 	// so will protect against function domain exceptions.  
 	//
 	if (indexBooleanChildren>0){
-		int finalSize = elements.size();
+		int finalSize = (int)elements.size();
 		int size = finalSize-startSize;
 		vector<StackElement>::reverse_iterator iter = elements.rbegin();
 		for (int offset = 0; offset < size; ++offset) {
